+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/source/BaseMediaSource.java
@@ -101,37 +101,46 @@ public abstract class BaseMediaSource implements MediaSource {
    */
   protected final MediaSourceEventListener.EventDispatcher createEventDispatcher(
       @Nullable MediaPeriodId mediaPeriodId) {
-    return eventDispatcher.withParameters(
-        /* windowIndex= */ 0, mediaPeriodId, /* mediaTimeOffsetMs= */ 0);
+    return eventDispatcher.withParameters(/* windowIndex= */ 0, mediaPeriodId);
   }
 
   /**
    * Returns a {@link MediaSourceEventListener.EventDispatcher} which dispatches all events to the
-   * registered listeners with the specified {@link MediaPeriodId} and time offset.
-   *
-   * @param mediaPeriodId The {@link MediaPeriodId} to be reported with the events.
-   * @param mediaTimeOffsetMs The offset to be added to all media times, in milliseconds.
-   * @return An event dispatcher with pre-configured media period id and time offset.
-   */
-  protected final MediaSourceEventListener.EventDispatcher createEventDispatcher(
-      MediaPeriodId mediaPeriodId, long mediaTimeOffsetMs) {
-    Assertions.checkNotNull(mediaPeriodId);
-    return eventDispatcher.withParameters(/* windowIndex= */ 0, mediaPeriodId, mediaTimeOffsetMs);
-  }
-
-  /**
-   * Returns a {@link MediaSourceEventListener.EventDispatcher} which dispatches all events to the
-   * registered listeners with the specified window index, {@link MediaPeriodId} and time offset.
+   * registered listeners with the specified window index and {@link MediaPeriodId}.
    *
    * @param windowIndex The timeline window index to be reported with the events.
    * @param mediaPeriodId The {@link MediaPeriodId} to be reported with the events. May be null, if
    *     the events do not belong to a specific media period.
-   * @param mediaTimeOffsetMs The offset to be added to all media times, in milliseconds.
-   * @return An event dispatcher with pre-configured media period id and time offset.
+   * @return An event dispatcher with pre-configured media period id.
    */
   protected final MediaSourceEventListener.EventDispatcher createEventDispatcher(
+      int windowIndex, @Nullable MediaPeriodId mediaPeriodId) {
+    return eventDispatcher.withParameters(windowIndex, mediaPeriodId);
+  }
+
+  /**
+   * Note: The {@code mediaTimeOffsetMs} passed to this method is ignored and not added to media
+   * times in any way.
+   *
+   * @deprecated Use {@link #createEventDispatcher(MediaPeriodId)} instead.
+   */
+  @Deprecated
+  protected final MediaSourceEventListener.EventDispatcher createEventDispatcher(
+      MediaPeriodId mediaPeriodId, long mediaTimeOffsetMs) {
+    Assertions.checkNotNull(mediaPeriodId);
+    return eventDispatcher.withParameters(/* windowIndex= */ 0, mediaPeriodId);
+  }
+
+  /**
+   * Note: The {@code mediaTimeOffsetMs} passed to this method is ignored and not added to media
+   * times in any way.
+   *
+   * @deprecated Use {@link #createEventDispatcher(int, MediaPeriodId)} instead.
+   */
+  @Deprecated
+  protected final MediaSourceEventListener.EventDispatcher createEventDispatcher(
       int windowIndex, @Nullable MediaPeriodId mediaPeriodId, long mediaTimeOffsetMs) {
-    return eventDispatcher.withParameters(windowIndex, mediaPeriodId, mediaTimeOffsetMs);
+    return eventDispatcher.withParameters(windowIndex, mediaPeriodId);
   }
 
   /**