+++ libraries/exoplayer/src/test/java/androidx/media3/exoplayer/source/MergingMediaPeriodTest.java
@@ -246,10 +246,7 @@ public final class MergingMediaPeriodTest {
           new FakeMediaPeriodWithSelectionParameters(
               new TrackGroupArray(trackGroups),
               new EventDispatcher()
-                  .withParameters(
-                      /* windowIndex= */ i,
-                      new MediaPeriodId(/* periodUid= */ i),
-                      /* mediaTimeOffsetMs= */ 0),
+                  .withParameters(/* windowIndex= */ i, new MediaPeriodId(/* periodUid= */ i)),
               /* trackDataFactory= */ (unusedFormat, unusedMediaPeriodId) ->
                   ImmutableList.of(
                       oneByteSample(definition.singleSampleTimeUs, C.BUFFER_FLAG_KEY_FRAME),